sync: Add synchronisation via fd & poll
Implement copy synchronisation via file description and pool. Every
memory object have fd which could be used in poll to check is the memory
ready.
First of all it provides clear and familiar way to use it, which leads
to easy integration in other projects (e.g. libevents). It also reduces
amount of extra objects and library complexity in general.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>